+++ akka-http-jwt/src/main/scala/akka/http/jwt/internal/JwtSettingsImpl.scala
@@ -1,4 +1,8 @@
-package akka.http.jwt.impl.settings
+/*
+ * Copyright (C) 2024 Lightbend Inc. <https://www.lightbend.com>
+ */
+
+package akka.http.jwt.internal
 
 import akka.annotation.InternalApi
 import akka.http.impl.util.SettingsCompanionImpl
