+++ akka-http-jwt/src/main/scala/akka/http/jwt/scaladsl/JwtSettings.scala
@@ -1,8 +1,12 @@
+/*
+ * Copyright (C) 2024 Lightbend Inc. <https://www.lightbend.com>
+ */
+
 package akka.http.jwt.scaladsl
 
 import akka.actor.ClassicActorSystemProvider
 import akka.annotation.{ ApiMayChange, DoNotInherit }
-import akka.http.jwt.impl.settings.{ JwtSettingsImpl, JwtSupport }
+import akka.http.jwt.internal.{JwtSettingsImpl, JwtSupport}
 import com.typesafe.config.Config
 
 @ApiMayChange @DoNotInherit
