+++ akka-http-jwt/src/main/scala/akka/http/jwt/scaladsl/server/directives/JwtDirectives.scala
@@ -1,17 +1,22 @@
-package akka.http.scaladsl.server.directives
+/*
+ * Copyright (C) 2024 Lightbend Inc. <https://www.lightbend.com>
+ */
+
+package akka.http.jwt.scaladsl.server.directives
 
 import akka.event.LoggingAdapter
-import akka.http.jwt.impl.settings.JwtSupport
+import akka.http.jwt.internal.JwtSupport
 import akka.http.jwt.scaladsl
 import akka.http.scaladsl.server.Directive1
 import akka.http.scaladsl.server.Directives.Authenticator
 import akka.http.scaladsl.server.Directives.authenticateOAuth2
-import spray.json.{ JsBoolean, JsNumber, JsObject, JsString }
+import akka.http.scaladsl.server.directives.Credentials
+import spray.json.{JsBoolean, JsNumber, JsObject, JsString}
 
 trait JwtDirectives {
 
-  import BasicDirectives._
-  import RouteDirectives._
+  import akka.http.scaladsl.server.directives.BasicDirectives._
+  import akka.http.scaladsl.server.directives.RouteDirectives._
 
   def jwt(): Directive1[JwtClaims] = {
     extractActorSystem.flatMap { system =>
@@ -37,23 +42,7 @@ trait JwtDirectives {
       }
   }
 
-  // JwtClaims provides utilities to easily assert and extract claims from the JWT token
-  class JwtClaims(claims: JsObject) {
 
-    def hasClaim(name: String): Boolean = claims.fields.contains(name)
-
-    def intClaim(name: String): Option[Int] = claims.fields.get(name).collect { case JsNumber(value) => value.toInt }
-
-    def longClaim(name: String): Option[Long] = claims.fields.get(name).collect { case JsNumber(value) => value.toLong }
-
-    def doubleClaim(name: String): Option[Double] = claims.fields.get(name).collect { case JsNumber(value) => value.toDouble }
-
-    def stringClaim(name: String): Option[String] = claims.fields.get(name).collect { case JsString(value) => value }
-
-    def booleanClaim(name: String): Option[Boolean] = claims.fields.get(name).collect { case JsBoolean(value) => value }
-
-    def toJson: String = claims.toString()
-  }
 
 }
